requirements.txt
google-cloud==0.34.0
numpy==1.24.4
pandas==2.0.3
openmeteo-requests==1.1.0
requests==2.31.0
requests-cache==1.1.1
retry-requests==2.0.0
scipy==1.10.1
